third_party/blink/web_tests/platform/generic/http/tests/security/local-video-source-from-remote-expected.txt

CONSOLE ERROR: Not allowed to load local resource: test.ogv
Test that a remote video element will not use a local <source>, and will use another remote <source>

This test only behaves correctly in DRT

EVENT(loadedmetadata)
SUCCESS: remote movie loaded, local movie failed to load

END OF TEST
